+++ projects/ng2-qgrid-theme-material/src/lib/components/date-adapter/format-date-adapter.ts
@@ -1,0 +1,15 @@
+import { DatePipe } from '@angular/common';
+import { NativeDateAdapter } from '@angular/material/core';
+
+export class FormatDateAdapter extends NativeDateAdapter {
+	useFormat: string;
+	datePipe: DatePipe;
+
+	format(date: Date, displayFormat: Object): string {
+		if (this.useFormat) {
+			return this.datePipe.transform(date, this.useFormat);
+		}
+
+		return super.format(date, displayFormat);
+	}
+}