+++ projects/ng2-qgrid/src/public-api.ts
@@ -143,6 +143,7 @@ export {
 	DataManipulationComponent,
 	DataManipulationModule,
 	DateDirective,
+	DateMaskDirective,
 	DateModule,
 	DatePipe,
 	EbClassDirective,
@@ -234,6 +235,7 @@ export { Shortcut } from '@qgrid/core/shortcut/shortcut';
 export { ShortcutDispatcher } from '@qgrid/core/shortcut/shortcut.dispatcher';
 export { CommandManager } from '@qgrid/core/command/command.manager';
 export { StyleCellContext, StyleRowContext } from '@qgrid/core/style/style.context';
+export { PersistenceStorage, Storage } from '@qgrid/core/persistence/persistence.storage';
 
 // @qgrid/core/columns
 export { ArrayColumnModel as ArrayColumn } from '@qgrid/core/column-type/array.column';