+++ projects/ngx-qgrid-plugins/src/lib/date/date.module.ts
@@ -1,12 +1,19 @@
 import { NgModule } from '@angular/core';
 import { DateDirective } from './date.directive';
+import { DateMaskDirective } from './date-mask.directive';
+import { DateService } from './date.service';
 
 @NgModule({
 	declarations: [
-		DateDirective
+		DateDirective,
+		DateMaskDirective,
 	],
 	exports: [
-		DateDirective
+		DateDirective,
+		DateMaskDirective
+	],
+	providers: [
+		DateService,
 	]
 })
 export class DateModule {