+++ projects/qgrid-core/command-bag/document.click.command.js
@@ -9,15 +9,14 @@ export class DocumentClickCommand extends Command {
         super({
             key: DOCUMENT_CLICK_COMMAND_KEY,
             execute: ([host, e]) => {
-                const path = eventPath(e);
-                const clickedOutside = path.every(el => el !== host);
-                if (clickedOutside) {
-                    switch(model.edit().status) {
-                        case 'cell': {
-                            const cancelEdit = commandPalette.get(EDIT_CELL_CANCEL_COMMAND_KEY);
-                            if(cancelEdit.canExecute() === true) {
-                                cancelEdit.execute();
-                            }
+                if (model.edit().status === 'edit') {
+                    const path = eventPath(e);
+                    const clickedOutside = path.every(x => x !== host && !x.classList.contains('q-grid-editor-part'));
+                    if (clickedOutside) {
+
+                        const cancelEdit = commandPalette.get(EDIT_CELL_CANCEL_COMMAND_KEY);
+                        if (cancelEdit.canExecute() === true) {
+                            cancelEdit.execute();
                         }
                     }
                 }