+++ projects/qgrid-core/command-bag/filter.row.commit.command.js
@@ -21,26 +21,36 @@ export class FilterRowCommitCommand extends Command {
                 if (!isUndefined(search) && search !== null && search !== '') {
                     const opList = operatorFactory(column);
                     const op = filter.expression ? filter.expression.op : opList[0];
-                    if (op === 'contains') {
-                        filter.items = [search];
-                    } else {
-                        filter.expression = {
-                            kind: 'condition',
-                            left: key,
-                            op,
-                            right: search
-                        };
+                    switch (op) {
+                        case 'contains': {
+                            filter.items = [search];
+                            break;
+                        }
+                        case 'between': {
+                            filter.expression = {
+                                kind: 'condition',
+                                left: key,
+                                op,
+                                right: [null, search]
+                            };
+                            break;
+                        }
+                        default: {
+                            filter.expression = {
+                                kind: 'condition',
+                                left: key,
+                                op,
+                                right: search
+                            };
+                            break;
+                        }
                     }
                 }
                 else {
                     delete by[key];
                 }
 
-                model.filter({
-                    by
-                }, {
-                    source: 'filter.let'
-                });
+                model.filter({ by }, { source: 'filter.let' });
 
                 return true;
             }