+++ projects/qgrid-core/expression/predicate.visitor.d.ts
@@ -1,5 +1,10 @@
 import { Visitor } from './expression.visitor';
+import { Assert } from '../filter/assert';
 
 export declare class PredicateVisitor extends Visitor {
-	constructor(valueFactory: (name: string) => (row: any) => any);
+	constructor(
+		valueFactory: (name: string) => (row: any) => any,
+		assertFactory: (name: string) => Assert,
+		getType: (name: string) => string,
+	);
 }