+++ projects/qgrid-core/keyboard/keyboard.d.ts
@@ -1,6 +1,6 @@
 export declare class Keyboard {
-	static isPrintable(key: string): boolean;
-	static isControl(key: string): boolean;
-	static stringify(key: string): string;
-	static translate(code: string): string;
+	static isPrintable(code: string): boolean;
+	static isControl(code: string): boolean;
+	static stringify(code: string, key: string): string;
+	static translate(code: number): string;
 }