+++ projects/qgrid-core/persistence/persistence.state.js
@@ -5,12 +5,14 @@ export class PersistenceState {
 	constructor() {
 		this.id = 'default';
 		this.defaultGroup = 'My Presets';
-		this.load = new Command();
-		this.remove = new Command();
-		this.create = new Command();
-		this.modify = new Command();
-		this.setDefault = new Command();
-		this.reset = new Command();
+		this.schedule = 'onDemand'; // onDemand | onStateChange
+
+		this.load = new Command({ source: 'persistence.model' });
+		this.remove = new Command({ source: 'persistence.model' });
+		this.create = new Command({ source: 'persistence.model' });
+		this.modify = new Command({ source: 'persistence.model' });
+		this.setDefault = new Command({ source: 'persistence.model' });
+		this.reset = new Command({ source: 'persistence.model' });
 
 		this.storage = new PersistenceStorage(localStorage);
 		this.settings = {
@@ -18,7 +20,8 @@ export class PersistenceState {
 			sort: ['by'],
 			pivot: ['by'],
 			filter: ['by'],
-			queryBuilder: ['node']
+			queryBuilder: ['node'],
+			pagination: ['current', 'size'],
 		};
 	}
 }