+++ CHANGELOG.md
@@ -9,8 +9,46 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Remove `options` property from $view.edit.cell.
 - Rename `shortcuts` to `shortcut` in the edit state.
 - Remove `prevent` property from the navigation state.
+- Remove `id` property from `Action` type.
 
-## [9.1.7] - 2020-07-20
+## [9.2.1] - 2020-10-19
+### Fixed
+- Fix datepicker which didn't work in edit mode correctly, by introducing `q-grid-editor-part` css class.
+- Empty date is a valid case.
+- Filter row works when between operator is chosen.
+
+### Added
+- Date mask to enhance user input performance.
+- Two new properties in datetime columns `dateFormat` and `timeFormat` to enhance editing.
+- Persistence state now accepts schedule equals to `onStateChange` that allows to auto save grid state on user changes.
+
+### Changed
+- Date input and datepicker takes the format from the date column by using DateAdapter.
+
+## [9.1.11] - 2020-08-19
+### Fixed
+- Do not focus grid on scene change.
+ 
+## [9.1.10] - 2020-08-17
+### Fixed
+- Dates are properly converted to midnights.
+- When first input date in datetime input time is set to midnight.
+
+## [9.1.9] - 2020-08-13
+### Fixed
+- Predicate visitor get type from the column and not resolve it by itself.
+- If start\end is not set using between operator it works like greaterOrEquals\lessOrEquals.
+
+## [9.1.8] - 2020-08-11
+### Fixed
+- Uncheck select all when switch to page with no-selection.
+
+## [9.1.7] - 2020-07-23
+### Fixed
+- Column filter between operator doesn't appear.
+- Column filter doesn't work properly due to the not valid filter value type.
+
+## [9.1.6] - 2020-07-20
 ### Fixed
 - Fix date editor width.
 - Fix scrolling when column sort.
@@ -23,12 +61,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - To remove column chooser it's not required to override toolbar template.
 
-## [9.1.6] - 2020-07-13
+## [9.1.5] - 2020-07-13
 ### Fixed
 - Importing of LIVR library.
 - Set compilation target to es5. 
 
-## [9.1.5] - 2020-06-02
 ### Added
 - Support 2 version 9.1.x - Ivy, 9.1.x-ve - View Engine.
 - `q-grid-shortcut-host` directive that can change key event source for q-grid commands.
@@ -573,9 +610,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Progress plugin.
 
 [unreleased]: https://github.com/qgrid/ng2/compare/v10.0.0...HEAD
-[10.0.0]: https://github.com/qgrid/ng2/compare/v9.1.7...v10.0.0
+[10.0.0]: https://github.com/qgrid/ng2/compare/v10.0.0...v9.2.1
+[9.2.1]: https://github.com/qgrid/ng2/compare/v9.1.11...v9.2.1
+[9.1.11]: https://github.com/qgrid/ng2/compare/v9.1.10...v9.1.11
+[9.1.10]: https://github.com/qgrid/ng2/compare/v9.1.9...v9.1.10
+[9.1.9]: https://github.com/qgrid/ng2/compare/v9.1.8...v9.1.9
+[9.1.8]: https://github.com/qgrid/ng2/compare/v9.1.7...v9.1.8
 [9.1.7]: https://github.com/qgrid/ng2/compare/v9.1.6...v9.1.7
-[9.1.6]: https://github.com/qgrid/ng2/compare/v9.1.4...v9.1.6
+[9.1.6]: https://github.com/qgrid/ng2/compare/v9.1.5...v9.1.6
+[9.1.5]: https://github.com/qgrid/ng2/compare/v9.1.4...v9.1.5
 [9.1.4]: https://github.com/qgrid/ng2/compare/v9.1.3...v9.1.4
 [9.1.3]: https://github.com/qgrid/ng2/compare/v9.0.1...v9.1.3
 [9.0.1]: https://github.com/qgrid/ng2/compare/v9.0.0...v9.0.1